Fix responding to blockwise requests using Write
When responding to blockwise requests using `ResponseWrite.Write` the
server would use the message ID of the first block message rather than
the last. This meant that clients would fail to handle the response.

This was due to not creating a new response writer when we created a new
request.