fix: do not attempt serve DTLS connection on accept error
Updates the DTLS server to skip serving a DTLS connection when an error
is returned. Previously, if an error was returned that did not meet the
criteria of shutting down the server altogether, a non-nil DTLS
connection would still attempt to be served. This can cause a panic if
the net.Conn interface is non-nil, but the underlying value is.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>