[udp] make response message cache configurable and update default
Modifies UDP conn implementation to allow for a custom response message
cache to be provided. This change is implemented in such a way that API
compatibility with previous v3 releases is preserved. Users may now
provide an option to supply their own cache implementation. Structured
messages, rather than their serialized representation, are now passed to
the cache implementation to allow for caching decisions to be made in
the cache implementation. For example, it may be desirable to skip
caching blockwise message responses if the entire underlying data being
transferred is also cached. The cache implementation is responsible for
cloning messages or otherwise ensuring that it is not storing data that
may subsequently be modified.

A slight functional change is also included in this update in that the
default cache implementation now has pseudo-LRU functionality via
ccache. When 30 newer messages are added to the cache, the 31st is
marked for deletion. Note that deletion is not immediate in ccache and
expired items can be returned. A check is added to ignore items that are
returned expired.

Further performance improvements can be made in the default cache
implementation. For example, while this implementation dramatically
improves memory consumption for connections performing large volumes of
requests, it also increases the total overhead of every new connection.
However, because the implementation is now pluggable, improvements can
be made without requiring updates to the library or breaking API
changes.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>